requirements.txt
# Production Dependencies
# Core dependencies for Kiwoom Trading System

# HTTP 클라이언트
requests>=2.31.0
websockets>=12.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 데이터 검증
pydantic>=2.5.0
pydantic-settings>=2.0.0

# AI/ML
google-generativeai>=0.3.0
lightgbm>=4.0.0
scikit-learn>=1.3.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0

# 시각화
matplotlib>=3.7.0
rich>=13.0.0

# 유틸리티
PyYAML>=6.0.0
python-telegram-bot>=20.0